@@ -5,7 +5,7 @@
   "license": "MIT",
 
   "require": {
-    "php": "^7.1|^8.0",
+    "php": "^8.2",
     "ext-openssl": "*",
     "ext-json": "*",
     "hidehalo/nanoid-php": "^1.1",
@@ -19,7 +19,10 @@
   "autoload": {
     "psr-4": {
       "Actinity\\SignedUrls\\": "src/"
-    }
+    },
+    "files": [
+      "src/helpers.php"
+    ]
   },
 
   "autoload-dev": {
